requirements.txt
openai>=1.0.0
edge-tts>=6.1.0
moviepy>=1.0.3
Pillow>=10.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pyyaml>=6.0
schedule>=1.2.0
requests>=2.31.0
replicate>=0.15.0

# Optional: better TTS voices (only if tts.provider=elevenlabs)
# elevenlabs>=1.0.0

# Optional: Prefect orchestration (only if orchestration.engine=prefect)
# prefect>=3.0.0
